version 2

+++ src/main/java/com/monster/nzway/domain/Company.java
@@ -1,5 +1,8 @@
 package com.monster.nzway.domain;
 
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableField;
+import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 
 /**
@@ -7,7 +10,8 @@ import lombok.Data;
  */
 @Data
 public class Company {
-    private Integer id;
+
+    private Long id;
     private String name;
     private String address;
     private String phone;
