+++ src/main/java/com/monster/nzway/domain/Distributor.java
@@ -1,5 +1,8 @@
 package com.monster.nzway.domain;
 
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableField;
+import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 
 /**
@@ -8,8 +11,12 @@ import lombok.Data;
 
 @Data
 public class Distributor {
-    private Integer id;
+
+    private Long id;
+
     private String name;
+
     private String address;
+
     private String phone;
 }
